requirement.txt
socketio
dash 
dash-socketio 
plotly
websockets